CQRS remove ibatis conext meaning. It is the node insert context.
