+++ src/main/java/org/alfresco/repo/domain/node/cqrs/CqrsService.java
@@ -24,12 +24,29 @@
  * #L%
  */
 
-package org.alfresco.repo.domain.node.ibatis.cqrs.utils;
+package org.alfresco.repo.domain.node.cqrs;
+
+import org.alfresco.repo.domain.node.cqrs.utils.Context;
 
 /**
+ * Service for implementing CQRS
+ *
  * Created by mmuller on 26/03/2018.
  */
-public interface CqrsLogger
+public interface CqrsService
 {
-    public void log(String string);
+    /**
+     * Commands are capable of changing the database. They will be evaluated and either accepted or refused.
+     * If they are refused they will not change the database.
+     *
+     * @param diffObject
+     */
+    public void executeCommand(Object diffObject);
+
+    /**
+     * Represents the context from which the service was called.
+     *
+     * @return context in which the service is called
+     */
+    public Context getContext();
 }
