+++ src/main/java/org/alfresco/repo/domain/node/cqrs/utils/CqrsLogger.java
@@ -24,20 +24,12 @@
  * #L%
  */
 
-package org.alfresco.repo.domain.node.ibatis.cqrs;
-
-import java.util.List;
+package org.alfresco.repo.domain.node.cqrs.utils;
 
 /**
- * Listener interface for Writer and Reader
- *
- * Created by mmuller on 27/03/2018.
+ * Created by mmuller on 26/03/2018.
  */
-public interface EventListener
+public interface CqrsLogger
 {
-    void onUpdate(List<Event> events);
-
-    void onCreate(List<Event> events);
-
-    void onDelete(List<Event> events);
+    public void log(String string);
 }
